backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-multipart==0.0.6
websockets==12.0
aiofiles==23.2.1

# Spark integration
pyspark==3.3.0
py4j==0.10.9.5

# Airflow client
apache-airflow-client==2.5.0
requests==2.31.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# MinIO
minio==7.2.0

# Docker integration
docker==6.1.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring
psutil==5.9.6

# Jupyter integration
jupyter-client==8.6.0
nbformat==5.9.2
nbconvert==7.11.0
